Habit Update
Users can now create premade habits or custom habits via the + add Habit button.

need to implement a category column in the habits table so that the program can distinguish custom habits, premade habits and which type of premade habits so that it can display everything properly. Also the starting habits for new profiles need to be updated.